misc fixes to s3 block storage and adding more tests

+++ src/pyoram/storage/boto3_s3_wrapper.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 
 try:
     import boto3
@@ -24,32 +25,41 @@ class Boto3S3Wrapper(object):
         self._s3 = boto3.session.Session(
             aws_access_key_id=aws_access_key_id,
             aws_secret_access_key=aws_secret_access_key,
-            region_name=region).resource('s3')
+            region_name=region_name).resource('s3')
         self._bucket = self._s3.Bucket(bucket_name)
 
-    def exists(self, name):
+    def exists(self, key):
         try:
-            self._bucket.Object(name).load()
+            self._bucket.Object(key).load()
         except botocore.exceptions.ClientError as e:
             if e.response['Error']['Code'] == "404":
-                return False
+                pass
             else:
                 raise e
         else:
             return True
+        # It's not a file. Check if it's a "directory".
+        for obj in self._bucket.objects.filter(
+                Prefix=key+"/"):
+            return True
+        return False
 
     def download(self, key):
-        return self._s3.meta.client.get_object(
-            Bucket=self._bucket.name,
-            Key=self._basename % key)['Body'].read()
+        try:
+            return self._s3.meta.client.get_object(
+                Bucket=self._bucket.name,
+                Key=key)['Body'].read()
+        except botocore.exceptions.ClientError:
+            raise IOError("Can not download key: %s"
+                          % (key))
 
     def upload(self, (key, block)):
         self._bucket.put_object(Key=key, Body=block)
 
-    def clear(self):
+    def clear(self, key):
         for obj in self._bucket.objects.filter(
-                Prefix=storage_name+"/"):
-            self._s3.Object(bucket.name, obj.key).delete()
+                Prefix=key+"/"):
+            self._s3.Object(self._bucket.name, obj.key).delete()
 
 class MockBoto3S3Wrapper(object):
     """
@@ -66,21 +76,39 @@ class MockBoto3S3Wrapper(object):
         self._bucket_name = os.path.abspath(
             os.path.normpath(bucket_name))
 
-    def exists(self, name):
+    # called within upload to create directory
+    # heirarchy on the fly
+    def _makedirs_if_needed(self, key):
+        if not os.path.exists(
+                os.path.dirname(
+                    os.path.join(self._bucket_name, key))):
+            os.makedirs(
+                os.path.dirname(
+                    os.path.join(self._bucket_name, key)))
+        assert not os.path.isdir(
+            os.path.join(self._bucket_name, key))
+
+    def exists(self, key):
         return os.path.exists(
-            os.path.join(self._bucket_name, name))
+            os.path.join(self._bucket_name, key))
 
     def download(self, key):
         with open(os.path.join(self._bucket_name, key), 'rb') as f:
             return f.read()
 
     def upload(self, (key, block)):
+        self._makedirs_if_needed(key)
         with open(os.path.join(self._bucket_name, key), 'wb') as f:
             f.write(block)
 
-    def clear(self, name):
-        shutil.rmtree(
-            os.path.join(self._bucket_name, name),
-            ignore_errors=True)
-        os.makedirs(
-            os.path.join(self._bucket_name, name))
+    def clear(self, key):
+        if os.path.exists(
+                os.path.join(self._bucket_name, key)):
+            if os.path.isdir(
+                    os.path.join(self._bucket_name, key)):
+                shutil.rmtree(
+                    os.path.join(self._bucket_name, key),
+                    ignore_errors=True)
+            else:
+                os.remove(
+                    os.path.join(self._bucket_name, key))
